requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-security
azure-mgmt-storage
azure-mgmt-monitor
azure-monitor-ingestion
azure-functions 
dotenv